requirements.txt
numpy==1.23.5
opencv-python==4.7.0.72
pylibtiff==0.5.1
pillow==9.4.0
nibabel==5.0.0
imageio==2.25.0
pandas==1.5.3
connected-components-3d==3.10.5
scipy==1.10.0
matplotlib==3.6.3


